migrated controller files for gtk4

+++ source/rafcon/gui/interface.py
@@ -17,13 +17,13 @@ from rafcon.core import interface as core_interface
 
 
 def add_library_root_path_to_shortcut_folders_of_dialog(dialog):
-    from gi.repository import GLib
+    from gi.repository import GLib, Gio
     from rafcon.gui.singleton import library_manager
     library_paths = library_manager.library_root_paths
     library_keys = sorted(library_paths)
     for library_key in library_keys:
         try:
-            dialog.add_shortcut_folder(library_paths[library_key])
+            dialog.add_shortcut_folder(Gio.File.new_for_path(library_paths[library_key]))
         except GLib.GError:
             # this occurs if the shortcut file already exists
             # unfortunately dialog.list_shortcut_folders() does not work
@@ -33,45 +33,40 @@ def add_library_root_path_to_shortcut_folders_of_dialog(dialog):
 
 def open_folder(query, default_path=None):
     """Shows a user dialog for folder selection
-    
-    A dialog is opened with the prompt `query`. The current path is set to the last path that was opened/created. The 
+
+    A dialog is opened with the prompt `query`. The current path is set to the last path that was opened/created. The
     roots of all libraries is added to the list of shortcut folders.
-    
+
     :param str query: Prompt asking the user for a specific folder
-    :param str default_path: Path to use if user does not specify one 
+    :param str default_path: Path to use if user does not specify one
     :return: Path selected by the user or `default_path` if no path was specified or None if none of the paths is valid
     :rtype: str
     """
-    from gi.repository import Gtk
-    from os.path import expanduser, pathsep, dirname, isdir
+    from gi.repository import Gtk, Gio
+    from os.path import expanduser, isdir
     from rafcon.gui.singleton import main_window_controller
     from rafcon.gui.runtime_config import global_runtime_config
+    from rafcon.gui.utils.dialog import run_dialog
     last_path = global_runtime_config.get_config_value('LAST_PATH_OPEN_SAVE', "")
-    selected_filename = None
-    if last_path and isdir(last_path):
-        selected_filename = last_path.split(pathsep)[-1]
-        last_path = dirname(last_path)
-    else:
-        last_path = expanduser('~')
 
     dialog = Gtk.FileChooserDialog(title=query, transient_for=None,
                                    action=Gtk.FileChooserAction.SELECT_FOLDER)
-    dialog.add_buttons(Gtk.STOCK_CANCEL, Gtk.ResponseType.CANCEL,
-                       Gtk.STOCK_OPEN, Gtk.ResponseType.OK)
+    dialog.add_buttons("_Cancel", Gtk.ResponseType.CANCEL,
+                       "_Open", Gtk.ResponseType.OK)
     # Allows confirming with Enter and double-click
     dialog.set_default_response(Gtk.ResponseType.OK)
     if main_window_controller:
         dialog.set_transient_for(main_window_controller.view.get_parent_widget())
-    dialog.set_current_folder(last_path)
-    if selected_filename is not None:
-        dialog.select_filename(selected_filename)
-
-    dialog.set_show_hidden(False)
+    if last_path and isdir(last_path):
+        # preselect the last opened folder
+        dialog.set_file(Gio.File.new_for_path(last_path))
+    else:
+        dialog.set_current_folder(Gio.File.new_for_path(expanduser('~')))
 
     # Add library roots to list of shortcut folders
     add_library_root_path_to_shortcut_folders_of_dialog(dialog)
 
-    response = dialog.run()
+    response = run_dialog(dialog)
 
     if response != Gtk.ResponseType.OK:
         dialog.destroy()
@@ -79,10 +74,11 @@ def open_folder(query, default_path=None):
             return default_path
         return None
 
-    path = dialog.get_filename()
+    selected_file = dialog.get_file()
+    path = selected_file.get_path() if selected_file else None
     dialog.destroy()
 
-    if os.path.isdir(path):
+    if path and os.path.isdir(path):
         global_runtime_config.set_config_value('LAST_PATH_OPEN_SAVE', path)
         return path
     return None
@@ -95,23 +91,24 @@ core_interface.open_folder_func = open_folder
 
 def create_folder(query, default_name=None, default_path=None, current_folder=None):
     """Shows a user dialog for folder creation
-    
-    A dialog is opened with the prompt `query`. The current path is set to the last path that was opened/created. The 
+
+    A dialog is opened with the prompt `query`. The current path is set to the last path that was opened/created. The
     roots of all libraries is added to the list of shortcut folders.
-    
+
     :param str query: Prompt asking the user for a specific folder
-    :param str default_name: Default name of the folder to be created 
+    :param str default_name: Default name of the folder to be created
     :param str default_path: Path in which the folder is created if the user doesn't specify a path
     :param str current_folder: Current folder that the FileChooserDialog points to in the beginning
     :return: Path created by the user or `default_path`/`default_name` if no path was specified or None if none of the
       paths is valid
     :rtype: str
     """
-    from gi.repository import Gtk
+    from gi.repository import Gtk, Gio
     from os.path import expanduser, dirname, join, exists, isdir
     from rafcon.core.storage.storage import STATEMACHINE_FILE
     from rafcon.gui.singleton import main_window_controller
     from rafcon.gui.runtime_config import global_runtime_config
+    from rafcon.gui.utils.dialog import run_dialog
     last_path = global_runtime_config.get_config_value('LAST_PATH_OPEN_SAVE', "")
 
     if last_path and isdir(last_path) and not exists(join(last_path, STATEMACHINE_FILE)):
@@ -123,19 +120,18 @@ def create_folder(query, default_name=None, default_path=None, current_folder=No
 
     dialog = Gtk.FileChooserDialog(title=query, transient_for=None,
                                    action=Gtk.FileChooserAction.CREATE_FOLDER)
-    dialog.add_buttons(Gtk.STOCK_CANCEL, Gtk.ResponseType.CANCEL,
-                       Gtk.STOCK_OPEN, Gtk.ResponseType.OK)
-    # dialog.add_buttons(Gtk.ButtonsType.OK_CANCEL)
+    dialog.add_buttons("_Cancel", Gtk.ResponseType.CANCEL,
+                       "_Open", Gtk.ResponseType.OK)
     # Allows confirming with Enter and double-click
     dialog.set_default_response(Gtk.ResponseType.OK)
     if main_window_controller:
         dialog.set_transient_for(main_window_controller.view.get_parent_widget())
-    dialog.set_current_folder(last_path)
-    if current_folder:
-        dialog.set_current_folder(current_folder)
+    if isdir(last_path):
+        dialog.set_current_folder(Gio.File.new_for_path(last_path))
+    if current_folder and isdir(current_folder):
+        dialog.set_current_folder(Gio.File.new_for_path(current_folder))
     if default_name:
         dialog.set_current_name(default_name)
-    dialog.set_show_hidden(False)
 
     # Add library roots to list of shortcut folders
     add_library_root_path_to_shortcut_folders_of_dialog(dialog)
@@ -143,7 +139,7 @@ def create_folder(query, default_name=None, default_path=None, current_folder=No
     # Run until the desired folder is found (warn if files are deleted)
     confirmed = 0
     while not confirmed:
-        response = dialog.run()
+        response = run_dialog(dialog)
 
         # Exit without saving
         if response != Gtk.ResponseType.OK:
@@ -154,13 +150,16 @@ def create_folder(query, default_name=None, default_path=None, current_folder=No
                     return default
             return None
 
-        path = dialog.get_filename()
+        selected_file = dialog.get_file()
+        path = selected_file.get_path() if selected_file else None
+        if path is None:
+            continue
 
         # Check if trying to save inside another state machine
         parent_folder_path = os.path.dirname(path)
         if 'statemachine.json' in os.listdir(parent_folder_path):
-            popup = Gtk.Dialog(title='Invalid state machine path', transient_for=dialog, flags=0)
-            popup.add_buttons(Gtk.STOCK_OK, Gtk.ResponseType.ACCEPT)
+            popup = Gtk.Dialog(title='Invalid state machine path', transient_for=dialog)
+            popup.add_buttons("_OK", Gtk.ResponseType.ACCEPT)
             text = 'You are trying to save the state machine inside another state machine:\n'\
                    '"{}"\n\n'\
                    'This would lead to an invalid library path structure.\n'\
@@ -171,26 +170,26 @@ def create_folder(query, default_name=None, default_path=None, current_folder=No
             label.set_margin_end(20)
             label.set_margin_top(15)
             label.set_margin_bottom(15)
-            popup.vbox.pack_start(label, True, True, 0)
-            label.show()
+            popup.get_content_area().append(label)
             popup.set_transient_for(dialog)
-            response = popup.run()
+            run_dialog(popup)
             popup.destroy()
 
             # Delete the folder that was created during selection
-            os.rmdir(path)
+            if os.path.isdir(path):
+                os.rmdir(path)
 
             # Set path to the parent of the state machine
             path = os.path.dirname(parent_folder_path)
-            dialog.set_current_folder(path)
+            dialog.set_current_folder(Gio.File.new_for_path(path))
             continue
 
         # Give a warning if the path already contains files
         files_in_path = os.listdir(path)
         if files_in_path:
             path_list = path.split('/')
-            popup = Gtk.Dialog(title='Path confirmation', transient_for=dialog, flags=0)
-            popup.add_buttons(Gtk.STOCK_CANCEL, Gtk.ResponseType.REJECT, Gtk.STOCK_OK, Gtk.ResponseType.ACCEPT)
+            popup = Gtk.Dialog(title='Path confirmation', transient_for=dialog)
+            popup.add_buttons("_Cancel", Gtk.ResponseType.REJECT, "_OK", Gtk.ResponseType.ACCEPT)
             text = 'Do you want to choose "{}" as a root folder?\n'\
                    'The {} file(s)/folder(s) inside will be deleted!'\
                    .format(path_list[-1], len(files_in_path))
@@ -199,15 +198,14 @@ def create_folder(query, default_name=None, default_path=None, current_folder=No
             label.set_margin_end(20)
             label.set_margin_top(15)
             label.set_margin_bottom(15)
-            popup.vbox.pack_start(label, True, True, 0)
-            label.show()
+            popup.get_content_area().append(label)
             popup.set_transient_for(dialog)
-            response = popup.run()
+            response = run_dialog(popup)
             popup.destroy()
 
             if not response == Gtk.ResponseType.ACCEPT:
                 path = '/'.join(path_list[:-1])
-                dialog.set_current_folder(path)
+                dialog.set_current_folder(Gio.File.new_for_path(path))
             else:
                 confirmed = 1
         else:
@@ -240,10 +238,11 @@ def save_folder(query, default_name=None):
     :rtype: str
     """
     from os.path import expanduser, dirname, join, exists, isdir
-    from gi.repository import Gtk
+    from gi.repository import Gtk, Gio
     from rafcon.core.storage.storage import STATEMACHINE_FILE
     from rafcon.gui.singleton import main_window_controller
     from rafcon.gui.runtime_config import global_runtime_config
+    from rafcon.gui.utils.dialog import run_dialog
     last_path = global_runtime_config.get_config_value('LAST_PATH_OPEN_SAVE', "")
 
     if last_path and isdir(last_path) and not exists(join(last_path, STATEMACHINE_FILE)):
@@ -255,31 +254,32 @@ def save_folder(query, default_name=None):
 
     dialog = Gtk.FileChooserDialog(title=query, transient_for=None,
                                    action=Gtk.FileChooserAction.SAVE)
-    dialog.add_buttons(Gtk.STOCK_CANCEL, Gtk.ResponseType.CANCEL,
-                       Gtk.STOCK_OPEN, Gtk.ResponseType.OK)
+    dialog.add_buttons("_Cancel", Gtk.ResponseType.CANCEL,
+                       "_Open", Gtk.ResponseType.OK)
     # Allows confirming with Enter and double-click
     dialog.set_default_response(Gtk.ResponseType.OK)
     if main_window_controller:
         dialog.set_transient_for(main_window_controller.view.get_parent_widget())
-    dialog.set_current_folder(last_path)
+    if isdir(last_path):
+        dialog.set_current_folder(Gio.File.new_for_path(last_path))
     if default_name:
         dialog.set_current_name(default_name)
-    dialog.set_show_hidden(False)
 
     # Add library roots to list of shortcut folders
     add_library_root_path_to_shortcut_folders_of_dialog(dialog)
 
-    response = dialog.run()
+    response = run_dialog(dialog)
 
     if response != Gtk.ResponseType.OK:
         dialog.destroy()
         return None
 
-    path = dialog.get_filename()
+    selected_file = dialog.get_file()
+    path = selected_file.get_path() if selected_file else None
     dialog.destroy()
 
     # check path existence
-    if not exists(dirname(path)):
+    if path is None or not exists(dirname(path)):
         return None
     return path
 
@@ -293,8 +293,9 @@ def show_notice(query, custom_buttons=None):
     from gi.repository import Gtk
     from rafcon.gui.helpers.label import set_button_children_size_request
     from rafcon.gui.singleton import main_window_controller
+    from rafcon.gui.utils.dialog import run_dialog
     from xml.sax.saxutils import escape
-    dialog = Gtk.MessageDialog(flags=Gtk.DialogFlags.MODAL, type=Gtk.MessageType.INFO, buttons=Gtk.ButtonsType.OK)
+    dialog = Gtk.MessageDialog(modal=True, message_type=Gtk.MessageType.INFO, buttons=Gtk.ButtonsType.OK)
     if custom_buttons is not None:
         for text, id in custom_buttons:
             dialog.add_button(text, id)
@@ -302,7 +303,7 @@ def show_notice(query, custom_buttons=None):
         dialog.set_transient_for(main_window_controller.view.get_parent_widget())
     dialog.set_markup(escape(query))
     set_button_children_size_request(dialog)
-    response = dialog.run()
+    response = run_dialog(dialog)
     dialog.destroy()
     return response
